Be sure volume is calculated as a size_t

+++ clover_hybrids/setup.c
@@ -75,7 +75,7 @@ int prompt=0,status;
     iseed=par_buf.iseed;
     
     number_of_nodes = numnodes();
-    volume=nx*ny*nz*nt;
+    volume=((size_t)nx)*ny*nz*nt;
     total_iters=0;
     return(prompt);
 }
