+++ ks_hl_spectrum/setup.c
@@ -61,7 +61,7 @@ int initial_set(){
   strcpy(job_id,par_buf.job_id);
   
   number_of_nodes = numnodes();
-  volume=nx*ny*nz*nt;
+  volume=((size_t)nx)*ny*nz*nt;
   total_iters=0;
   return(prompt);
 }
